+++ Runtime/FloatSet.cs
@@ -5,7 +5,8 @@ namespace UnderLogic.Collections
     [CreateAssetMenu(menuName = "Collections/Sets/Float Set")]
     public class FloatSet : RuntimeSet<float>
     {
-        public void Add(int value) => Add((float)value);
-        public bool Remove(int value) => Remove((float)value);
+        public void Add(int value) => base.Add(value);
+        public bool TryAdd(int value) => base.TryAdd(value);
+        public bool Remove(int value) => base.Remove(value);
     }
 }